Separate dataset and status page clients
